Organized angular file structure
