+++ src/main/web-app/src/app/app-routing.module.ts
@@ -5,8 +5,8 @@ import { BaseViewComponent } from './base-view/base-view.component';
 import { RequestComponent } from './request/request.component';
 import { ManagerDashboardComponent } from './manager-dashboard/manager-dashboard.component';
 import { LoginComponent } from './login/login.component';
-import { AuthGuard } from './auth.guard';
-import { ManagerGuard } from './manager.guard';
+import { AuthGuard } from './auth-guards/auth.guard';
+import { ManagerGuard } from './auth-guards/manager.guard';
 
 
 const routes: Routes = [
