+++ src/main/web-app/src/app/app.module.ts
@@ -15,12 +15,12 @@ import { MaterialModules } from './material';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
 import 'hammerjs';
-import { GenericDialogComponent } from './generic-dialog/generic-dialog.component';
-import { DisplayDataDialogComponent } from './display-data-dialog/display-data-dialog.component';
-import { RequestDialogComponent } from './request-dialog/request-dialog.component';
+import { GenericDialogComponent } from './dialog-boxes/generic-dialog/generic-dialog.component';
+import { DisplayDataDialogComponent } from './dialog-boxes/display-data-dialog/display-data-dialog.component';
+import { RequestDialogComponent } from './dialog-boxes/request-dialog/request-dialog.component';
 import { LoginComponent } from './login/login.component';
-import { DiscussionDialogComponent } from './discussion-dialog/discussion-dialog.component';
-import {TokenInterceptorService} from './token-interceptor.service';
+import { DiscussionDialogComponent } from './dialog-boxes/discussion-dialog/discussion-dialog.component';
+import {TokenInterceptorService} from './services/token-interceptor.service';
