+++ src/main/web-app/src/app/auth-guards/auth.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router,RouterStateSnapshot,ActivatedRouteSnapshot} from '@angular/router';
-import {AuthService} from './auth.service';
+import {AuthService} from '../services/auth.service';
 
 @Injectable({
   providedIn: 'root'
