+++ src/main/web-app/src/app/auth-guards/manager.guard.spec.ts
