+++ src/main/web-app/src/app/dialog-boxes/discussion-dialog/discussion-dialog.component.html
