+++ src/main/web-app/src/app/dialog-boxes/display-data-dialog/display-data-dialog.component.html
