+++ src/main/web-app/src/app/dialog-boxes/generic-dialog/generic-dialog.component.ts
