+++ src/main/web-app/src/app/services/auth.service.spec.ts
